Move Invoices\PaymentMethod to Payments\PaymentMethod
